requirements.txt
Flask
feedparser
gunicorn
elevenlabs
python-dotenv
requests